rename my menu tab files for consistency
