+++ client/src/pages/account/MyMenusTab.js
@@ -3,7 +3,7 @@ import axios from "axios";
 import {FaCheck, FaPlus} from "react-icons/fa";
 import {FaX} from "react-icons/fa6";
 import MenuCard from "../../components/Admin/MenuCard";
-import "../../format/MenusTab.css"
+import "../../format/MyMenusTab.css"
 
 const MenusTab = ({user}) => {
